Small fixes on the chat process (#11)
*tweakings and improvement on oobabooga parameters (wip, need settings page improvement too)

+++ src/AIdentities.Shared/Plugins/Connectors/Completion/DefaultCompletionRequest.cs
@@ -7,14 +7,14 @@ public class DefaultCompletionRequest : ICompletionRequest
    public int? CompletionResults { get; init; }
    public IList<string>? StopSequences { get; init; }
    public string? UserId { get; init; }
-   public decimal? Temperature { get; init; }
-   public decimal? RepetitionPenality { get; init; }
-   public decimal? RepetitionPenalityRange { get; init; }
-   public decimal? TopPSamplings { get; init; }
-   public decimal? TopKSamplings { get; init; }
-   public decimal? TopASamplings { get; init; }
-   public decimal? TypicalSampling { get; init; }
-   public decimal? TailFreeSampling { get; init; }
+   public float? Temperature { get; init; }
+   public float? RepetitionPenality { get; init; }
+   public float? RepetitionPenalityRange { get; init; }
+   public float? TopPSamplings { get; init; }
+   public float? TopKSamplings { get; init; }
+   public float? TopASamplings { get; init; }
+   public float? TypicalSampling { get; init; }
+   public float? TailFreeSampling { get; init; }
    public int? MaxGeneratedTokens { get; init; }
    public int? ContextSize { get; init; }
    public object? LogitBias { get; init; }
